requirements.txt
Django==1.6.2
South==0.8.4
dj-static==0.0.5
django-auth-ldap==1.2.0
django-cors-headers==0.12
django-rest-swagger==0.1.14
djangorestframework==2.3.13
psycopg2==2.5.2
pystache==0.5.4
python-ldap==2.4.15
requests==2.2.1
static==1.0.2
uWSGI==2.0.4
wsgiref==0.1.2
